Refactor private project git methods into abstract base class
